Add a CanonicalOSSALifetime utility.
This was originally part of the CopyPropagation prototype. Now OSSA is
specified completely enough that it can be turned into a simple
utility.

PrunedLiveness is an completely general purpose lifetime analysis at
the CFG level, where the client only wants to track the live
blocks. It doesn't need to be tied to a particular SILValue.

CanonicalOSSALifetime uses PrunedLiveness to find the extended live
range and identify the consumes on the boundary. All other consumes
need their own copy. No other copies are needed.

By running this after other transformations that affect OSSA
lifetimes, we can avoid the need to indefinitely build ad-hoc
optimizations in SemanticARC to recover from suboptimal patterns.

TODO: handle guaranteed values. This should be a very straightforward
extension.