Enable MandatoryCopyPropagation at -Onone
This won't be quite as powerful as the -O form of the pass because it
needs to respect debug information (this version of the pass won't
remove debug_value instructions).

There are two reasons to do this, both related to predictable behavior
across Debug and Release builds.

1. Shortening object lifetimes can cause observable changes in program
   behavior in the presense of weak/unowned reference and
   deinitializer side effects.

2. Programmers need to know reliably whether a given code pattern will
   copy the storage for copy-on-write types (Array, Set). Eliminating
   the "unexpected" copies the same way at -Onone and -O both makes
   debugging tractable and provides assurance that the code isn't
   relying on the luck of the optimizer in a particular compiler
   release.

Fixes rdar://73903961 (Enable CopyPropagation at -Onone)