Add insertOSSADestroys utility and fix PredMemOpt
PredictableMemOpts is buggy because it doesn't consider that there may
be a non-consuming use in a block where the consuming use was already
deleted.

Replace it with a much simpler insertOSSADestroys utility.

Problem: this utility cannot handle reborrows yet. To do that, we also
need to insert copied. I already have a utility on a different branch
that does that.