Add support to OwnershipRAUWHelper for terminator results.
The 'oldValue' can now be a terminator result instead of a SingleValueInstruction.

Plus fix some random comment typos in this utility.

Added a big urgent FIXME!!! for the sibling "single value"
utility. Fixing that is out of scope because the solution may look
different after other changes land. Just didn't want to forget about
it.